f32 migration: bulk f64→f32 + proto-boundary casts + loosened tolerances
Convert every f64 in session_rust/src/ to f32 so vertex data can be
bytemuck::cast_slice'd zero-copy into wgpu vertex buffers. wgpu has no
Float64x3 vertex format on most backends.

Wire format (session_proto/*.proto) stays double for cross-language
compat with C++/Python; casts are added at every prost boundary.

Tolerance constants loosened for f32 precision (epsilon ≈ 1.2e-7):
  ABSOLUTE        1e-9  → 1e-5
  RELATIVE        1e-6  → 1e-5
  ANGULAR         1e-6  → 1e-5
  ZERO_TOLERANCE  1e-12 → 1e-7

Sentinel infinities 1e308 / 1e300_f32 replaced with f32::INFINITY where
they exceeded f32 range.

659/673 minitests pass (97.9%). Remaining 14 failures: tolerance test
rebaselines (2), expected NURBS precision drift (7), algorithm-level
issues to investigate (4), Vector magnitude string format (1).

+++ examples/test_grid.rs
@@ -42,7 +42,7 @@ fn main() {
     println!("First 10 points (row 0):");
     for j in 0..10 {
         let u = u_min + (u_max - u_min) * 0.0 / 9.0;
-        let v = v_min + (v_max - v_min) * (j as f64) / 9.0;
+        let v = v_min + (v_max - v_min) * (j as f32) / 9.0;
         let pt = srf.point_at(u, v).unwrap();
         println!("({:.3}, {:.3}, {:.3})", pt[0], pt[1], pt[2]);
     }
@@ -50,7 +50,7 @@ fn main() {
     println!("\nNext 10 points (row 1):");
     for j in 0..10 {
         let u = u_min + (u_max - u_min) * 1.0 / 9.0;
-        let v = v_min + (v_max - v_min) * (j as f64) / 9.0;
+        let v = v_min + (v_max - v_min) * (j as f32) / 9.0;
         let pt = srf.point_at(u, v).unwrap();
         println!("({:.3}, {:.3}, {:.3})", pt[0], pt[1], pt[2]);
     }
@@ -58,7 +58,7 @@ fn main() {
     println!("\nRow 2 points:");
     for j in 0..10 {
         let u = u_min + (u_max - u_min) * 2.0 / 9.0;
-        let v = v_min + (v_max - v_min) * (j as f64) / 9.0;
+        let v = v_min + (v_max - v_min) * (j as f32) / 9.0;
         let pt = srf.point_at(u, v).unwrap();
         println!("({:.3}, {:.3}, {:.3})", pt[0], pt[1], pt[2]);
     }
