+++ src/nurbscurve_test.rs
@@ -103,7 +103,7 @@ pub fn run_nurbscurve_create_fitted() -> TestResult {
 
         // Open: 21 points on sine wave → fit with 8 CVs
         let pts: Vec<Point> = (0..=20).map(|i| {
-            let t = i as f64 * 2.0 * PI / 20.0;
+            let t = i as f32 * 2.0 * PI / 20.0;
             Point::new(t, 3.0 * t.sin(), 0.0)
         }).collect();
 
@@ -118,7 +118,7 @@ pub fn run_nurbscurve_create_fitted() -> TestResult {
 
         // Periodic: 24 points on circle → fit with 10 free CVs
         let cpts: Vec<Point> = (0..24).map(|i| {
-            let a = i as f64 * 2.0 * PI / 24.0;
+            let a = i as f32 * 2.0 * PI / 24.0;
             Point::new(a.cos(), a.sin(), 0.0)
         }).collect();
 
@@ -613,7 +613,7 @@ pub fn run_nurbscurve_modifications() -> TestResult {
         }
 
         for i in 0..curve_open.nurbsknot_count() {
-            curve_open.set_nurbsknot(i, i as f64 * 1.0);
+            curve_open.set_nurbsknot(i, i as f32 * 1.0);
         }
 
         // Now clamp, making 2 nurbsknots at the ends the same
